Add filter orgs by tag functionality  #215, remove extraneous filtering of challenge cards #19 #20 #192
