+++ server/handlers/all-users-view.js
@@ -11,7 +11,7 @@ module.exports = (request, reply) => {
       Hoek.assert(!error, 'redis error');
       var allUsers = attachOrgsToUsers(stringifiedOrgs, stringifiedUsers);
       var users = {
-        allUsers: loggedIn.scope === 'admin' ? allUsers : sortAlphabetically(filterActiveAndAdmin(allUsers)),
+        allUsers: loggedIn.scope === 'admin' ? allUsers : helpers.sortAlphabetically('first_name')(filterActiveAndAdmin(allUsers)),
         alternate: [{
           path: '/people/add',
           name: '+'
@@ -28,25 +28,6 @@ module.exports = (request, reply) => {
 
 function filterActiveAndAdmin (arr) {
   return arr.filter(el => el.active && el.user_type !== 'admin');
-}
-
-// TODO: put cloneArry in helpers
-function cloneArray (arr) {
-  return arr.map(el => Object.assign({}, el));
-}
-
-function sortAlphabetically (arr) {
-  return cloneArray(arr).sort((user1, user2) => {
-    var name1 = user1.first_name.toUpperCase();
-    var name2 = user2.first_name.toUpperCase();
-    if (name1 < name2) {
-      return -1;
-    }
-    if (name1 > name2) {
-      return 1;
-    }
-    return 0;
-  });
 }
 
 function attachOrgsToUsers (stringifiedOrgs, stringifiedUsers) {
