+++ templates/partials/challenge.html
@@ -6,8 +6,8 @@
   {{/if}}
   <p class="card__header card__header--challenge">{{title}}</p>
   <p class="card__copy">{{description}}</p>
-  {{#if tags}}
-    {{#each tags}}
+  {{#if tagsData}}
+    {{#each tagsData}}
       <span class="tag__label">{{name}}</span>
     {{/each}}
   {{/if}}
