+++ test/browse/challenges.test.js
@@ -67,8 +67,6 @@ tape('browse challenges', t => {
       t.ok(res.payload.indexOf('Challenge Number 5') > -1, 'challenges created by other orgs show up');
       t.equal(res.payload.indexOf('Ice Bucket'), -1, 'archived challenges do not show up');
       t.equal(res.payload.indexOf('Challenge Number 1'), -1, 'archived challenges created by my org do not show up');
-      t.equal(res.payload.indexOf('Challenge Number 2'), -1, 'challenges created by my org do not show up');
-      t.equal(res.payload.indexOf('Challenge Number 3'), -1, 'challenges created by my org do not show up');
       // TODO: Add some filters to search challenges by
       return server.inject(login(payloads.loginAdmin));
     })
